+++ lib/utils/regex_collection.dart
@@ -63,44 +63,78 @@ class RegexCollection {
   /// except this will catch all other remaining characters
   static const except = r'[^\w\x00+-/*<>=&|\^%?@!~\.]+';
 
+  static final regExpTitle = RegExp(title);
+
+  static final regExpNumber = RegExp(number);
+
+  static final regExpAnyString = RegExp(anyString);
+
+  static final regExpNullChar = RegExp(nullChar);
+
+  static final regExpIdentifier = RegExp(identifier);
+
+  static final regExpOperators = RegExp(operators);
+
+  static final regExpAnyComment = RegExp([
+    hashComment,
+    doubleSlashComment,
+    multilineComment,
+    xmlComment
+  ].join('|'));
+
+  static final regExpSingleLineComment =
+      RegExp([hashComment, doubleSlashComment].join("|"));
+
+  static final regExpPrivate = RegExp(private);
+
+  static final regExpStringOrComment = RegExp([
+    backtickString,
+    tripleSingleQuoteString,
+    tripleDoubleQuoteString,
+    singleQuoteString,
+    doubleQuoteString,
+    hashComment,
+    doubleSlashComment,
+    multilineComment,
+    xmlComment
+  ].join('|'));
+
+  static final regExpTokenizer =
+      RegExp([nullChar, identifier, number, operators, except].join('|'));
+
   // functions for checking
 
   /// returns true if input token is an identifier and starts with capital letter
-  static bool isTitle(String token) =>
-      RegExp(title).stringMatch(token) == token;
+  static bool isTitle(String token) => regExpTitle.stringMatch(token) == token;
 
   /// returns true if input token is of type number
   static bool isNumber(String token) =>
-      RegExp(number).stringMatch(token) == token;
+      regExpNumber.stringMatch(token) == token;
 
   /// returns true if input token is of type string
   static bool isString(String token) =>
-      RegExp(anyString).stringMatch(token) == token;
+      regExpAnyString.stringMatch(token) == token;
 
   /// returns true if input token is a null character (i.e. \x00)
-  static bool isNullChar(String token) => RegExp(nullChar).hasMatch(token);
+  static bool isNullChar(String token) => regExpNullChar.hasMatch(token);
 
   /// returns true if token is an identifier
   static bool isIdentifier(String token) =>
-      RegExp(identifier).stringMatch(token) == token;
+      regExpIdentifier.stringMatch(token) == token;
 
   /// returns true if input token is an operator
   static bool isOperator(String token) =>
-      RegExp(operators).stringMatch(token) == token;
+      regExpOperators.stringMatch(token) == token;
 
   /// returns true if input matches regex of any comment single line or multiline
   static bool isComment(String token) =>
-      RegExp([hashComment, doubleSlashComment, multilineComment, xmlComment]
-              .join('|'))
-          .stringMatch(token) ==
-      token;
+      regExpAnyComment.stringMatch(token) == token;
 
   /// returns true if input is a single line comment
   static bool isSingleLineComment(String token) =>
-      RegExp([hashComment, doubleSlashComment].join("|")).stringMatch(token) ==
-      token;
+      regExpSingleLineComment.stringMatch(token) == token;
 
   /// returns true if input token is an identifier and starts with underscore(_)
   static bool isPrivate(String token) =>
-      RegExp(private).stringMatch(token) == token;
+      regExpPrivate.stringMatch(token) == token;
 }