+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## 0.0.3
+## 0.0.4
 
-* line numbers issues fixed
-* documentation updated
+* removed default ScrollController
+* documentation added to any_syntax_highlighter.dart
+* optimizations
+* fixed overflow error when lineNumbers is set to true